+++ ksconf/cli.py
@@ -23,7 +23,7 @@ from typing import List
 
 import ksconf._ksconf as ksc
 import ksconf.util
-from ksconf.commands import DescriptionHelpFormatterPreserveLayout, get_all_ksconf_cmds
+from ksconf.command import DescriptionHelpFormatterPreserveLayout, get_all_ksconf_cmds
 from ksconf.consts import EXIT_CODE_ENV_BUSTED, EXIT_CODE_INTERNAL_ERROR, KSCONF_DEBUG, is_debug
 from ksconf.hook import plugin_manager
 from ksconf.util.completers import autocomplete