+++ ksconf/commands/merge.py
@@ -12,7 +12,7 @@ from __future__ import absolute_import, unicode_literals
 
 import os
 
-from ksconf.commands import ConfFileProxy, ConfFileType, KsconfCmd, dedent
+from ksconf.command import ConfFileProxy, ConfFileType, KsconfCmd, dedent
 from ksconf.conf.merge import merge_conf_files
 from ksconf.conf.parser import PARSECONF_MID, PARSECONF_STRICT
 from ksconf.consts import EXIT_CODE_BAD_ARGS, EXIT_CODE_SUCCESS